# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/media": {
      "post": {
        "operationId": "MediaController_upload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Media file to upload",
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "The media file (image or video)"
                  },
                  "postId": {
                    "type": "number",
                    "description": "Optional: ID of the post to associate with"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Media uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Upload a media file (image or video) to Cloudinary",
        "tags": [
          "Media"
        ]
      },
      "get": {
        "operationId": "MediaController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of media files",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MediaResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "List all uploaded media",
        "tags": [
          "Media"
        ]
      }
    },
    "/media/{id}": {
      "delete": {
        "operationId": "MediaController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Media deleted successfully"
          },
          "404": {
            "description": "Media not found"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Delete a media file from Cloudinary and database",
        "tags": [
          "Media"
        ]
      }
    },
...
    "/moderation/text": {
      "post": {
        "operationId": "ModerationController_moderateText",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModerateTextDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Moderation result returned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModerationResultResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Moderate Arabic text",
        "tags": [
          "Moderation"
        ]
      }
    },
    "/moderation/image": {
      "post": {
        "operationId": "ModerationController_moderateImage",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModerateImageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Moderation result returned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModerationResultResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Moderate an image by URL",
        "tags": [
          "Moderation"
        ]
      }
    },
    "/moderation/video": {
      "post": {
        "operationId": "ModerationController_moderateVideo",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModerateVideoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Moderation result returned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModerationResultResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Moderate a video by URL",
        "tags": [
          "Moderation"
        ]
      }
    },
    "/moderation/reports": {
      "get": {
        "operationId": "ModerationController_getReports",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of moderation reports",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ModerationResultResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Fetch moderation reports",
        "tags": [
          "Moderation"
        ]
      }
    },
...
    "/notifications": {
      "get": {
        "operationId": "NotificationsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NotificationResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "List notifications for current user",
        "tags": [
          "Notifications"
        ]
      }
    },
    "/notifications/{notificationId}/read": {
      "patch": {
        "operationId": "NotificationsController_markRead",
        "parameters": [
          {
            "name": "notificationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Mark notification as read",
        "tags": [
          "Notifications"
        ]
      }
    },
    "/notifications/unread-count": {
      "get": {
        "operationId": "NotificationsController_unreadCount",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CountDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Get unread notifications count",
        "tags": [
          "Notifications"
        ]
      }
    },
    "/notifications/mark-all-read": {
      "patch": {
        "operationId": "NotificationsController_markAllRead",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Mark all notifications as read",
        "tags": [
          "Notifications"
        ]
      }
    },
...
  },
  "info": {
    "title": "YSHAI API",
    "description": "Generated for Orval client",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken"
      }
    },
    "schemas": {
...
      "MediaResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "url": {
            "type": "string",
            "example": "https://res.cloudinary.com/your-cloud/media/upload/v123/post_1.jpg"
          },
          "type": {
            "type": "string",
            "enum": [
              "image",
              "video"
            ],
            "example": "image"
          },
          "orderIndex": {
            "type": "number",
            "example": 0
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-08-22T10:00:00Z"
          },
          "postId": {
            "type": "number",
            "example": 123,
            "description": "Associated post ID (if any)"
          }
        },
        "required": [
          "id",
          "url",
          "type",
          "orderIndex",
          "createdAt",
          "postId"
        ]
      },
...
      "MessageDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
...
      "ModerateTextDto": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "description": "Arabic text to moderate",
            "maxLength": 5000,
            "example": "اشتري الآن واحصل على خصم 50%"
          },
          "postId": {
            "type": "number",
            "description": "Related post ID",
            "example": 123
          }
        },
        "required": [
          "text",
          "postId"
        ]
      },
      "ModerationResultResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "provider": {
            "type": "string",
            "enum": [
              "gemini",
              "custom"
            ]
          },
          "verdict": {
            "type": "string",
            "enum": [
              "allowed",
              "flagged",
              "blocked"
            ]
          },
          "details": {
            "type": "object",
            "nullable": true
          },
          "checkedAt": {
            "format": "date-time",
            "type": "string"
          },
          "postId": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "provider",
          "verdict",
          "checkedAt",
          "postId"
        ]
      },
      "ModerateImageDto": {
        "type": "object",
        "properties": {
          "imageUrl": {
            "type": "string",
            "description": "Publicly accessible image URL to moderate",
            "example": "https://example.com/image.jpg"
          },
          "postId": {
            "type": "number",
            "description": "Related post ID",
            "example": 123
          }
        },
        "required": [
          "imageUrl",
          "postId"
        ]
      },
      "ModerateVideoDto": {
        "type": "object",
        "properties": {
          "videoUrl": {
            "type": "string",
            "description": "Publicly accessible video URL to moderate",
            "example": "https://example.com/video.mp4"
          },
          "postId": {
            "type": "number",
            "description": "Related post ID",
            "example": 123
          }
        },
        "required": [
          "videoUrl",
          "postId"
        ]
      },
...
      "NotificationResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "type": {
            "type": "string",
            "enum": [
              "post_scheduled",
              "publish_failed",
              "ai_ready",
              "approved"
            ]
          },
          "payload": {
            "type": "object"
          },
          "read": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "link": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "id",
          "type",
          "payload",
          "read",
          "createdAt"
        ]
      },
      "CountDto": {
        "type": "object",
        "properties": {
          "count": {
            "type": "number"
          }
        },
        "required": [
          "count"
        ]
      },
...
    }
  }
}